Split by coords with collision to plane

+++ src/cloth.h
@@ -58,12 +58,12 @@ struct Cloth {
   void buildClothMesh();
   void split_cloth(vector<Cloth*> &cloth_objects);
 
-  void split_cloth_by_coord(vector<Cloth*> &cloth_objects, std::set<int> &slice_coords_set);
+  void split_cloth_by_coord(vector<Cloth*> &cloth_objects, std::set<float> &slice_coords_set);
   void create_cloth(double width, double height, int num_width_points, int num_height_points, double thickness, e_orientation orientation, vector<Cloth*> &cloth_objects,  vector<PointMass > pms);
   void build_springs();
 
   void build_spatial_map();
-  void build_bucket_map(set<int> &slice_coords_set);
+  void build_bucket_map(set<float> &slice_coords_set);
   void self_collide(PointMass &pm, double simulation_steps);
   float hash_position(Vector3D pos);
